# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification","description":"OpenApi documentation for Spring Hotel Management System","termsOfService":"Terms of service","contact":{"name":"Alaa Darawish,Laith Hamaile ","url":"https://aloshcoding.com","email":"alosh@mail.com"},"license":{"name":"Licence name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Local ENV"},{"url":"https://some-url.com","description":"HOTEL ENV"}],"security":[{"bearerAuth":[]}],"paths":{"/api/v1/rooms/{id}":{"get":{"tags":["room-controller"],"operationId":"getRoomById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}}},"put":{"tags":["room-controller"],"operationId":"updateRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}}},"delete":{"tags":["room-controller"],"operationId":"deleteRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/reservations/{id}":{"put":{"tags":["reservation-controller"],"operationId":"updateReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["reservation-controller"],"operationId":"cancelReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/v1/invoices/{id}":{"get":{"tags":["invoice-controller"],"operationId":"getInvoiceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceResponse"}}}}}},"put":{"tags":["invoice-controller"],"operationId":"updateInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceResponse"}}}}}},"delete":{"tags":["invoice-controller"],"operationId":"deleteInvoice","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/housekeeping-tasks/{id}":{"get":{"tags":["housekeeping-task-controller"],"operationId":"getTaskById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HousekeepingTaskResponse"}}}}}},"put":{"tags":["housekeeping-task-controller"],"operationId":"updateTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HousekeepingTaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HousekeepingTaskResponse"}}}}}},"delete":{"tags":["housekeeping-task-controller"],"operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/employees/{id}":{"get":{"tags":["employee-controller"],"operationId":"getEmployeeById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}},"put":{"tags":["employee-controller"],"operationId":"updateEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/api/v1/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getCustomerProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}},"put":{"tags":["customer-controller"],"operationId":"updateCustomerProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/admin/update/{id}":{"put":{"tags":["admin-controller"],"operationId":"updateAdmine","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}}}}}},"/api/v1/rooms":{"get":{"tags":["room-controller"],"operationId":"getAllRooms","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomResponse"}}}}}}},"post":{"tags":["room-controller"],"operationId":"addRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomResponse"}}}}}}},"/api/v1/reservations":{"post":{"tags":["reservation-controller"],"operationId":"reserveRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/invoices":{"get":{"tags":["invoice-controller"],"operationId":"getAllInvoices","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InvoiceResponse"}}}}}}},"post":{"tags":["invoice-controller"],"operationId":"createInvoice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InvoiceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InvoiceResponse"}}}}}}},"/api/v1/housekeeping-tasks":{"get":{"tags":["housekeeping-task-controller"],"operationId":"getAllTasks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HousekeepingTaskResponse"}}}}}}},"post":{"tags":["housekeeping-task-controller"],"operationId":"createTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HousekeepingTaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HousekeepingTaskResponse"}}}}}}},"/api/v1/employees":{"get":{"tags":["employee-controller"],"operationId":"getAllEmployees","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Employee"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"addEmployee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Employee"}}}}}}},"/api/v1/customers/register":{"post":{"tags":["customer-controller"],"operationId":"registerCustomer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/customers/change-password/{id}":{"post":{"tags":["customer-controller"],"operationId":"changePassword","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["authentication-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/auth/refresh-token":{"post":{"tags":["authentication-controller"],"operationId":"refreshToken","responses":{"200":{"description":"OK"}}}},"/api/v1/auth/authenticate":{"post":{"tags":["authentication-controller"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/v1/admin/login":{"post":{"tags":["admin-controller"],"operationId":"logIn","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/admin/add":{"post":{"tags":["admin-controller"],"operationId":"addAdmine","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}}}}}},"/api/v1/users":{"patch":{"tags":["user-controller"],"operationId":"changePassword_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/search/reservations":{"get":{"tags":["search-controller"],"operationId":"searchReservations","parameters":[{"name":"customerName","in":"query","required":false,"schema":{"type":"string"}},{"name":"customerId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/search/customer/{id}":{"get":{"tags":["search-controller"],"operationId":"searchCustomer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/search/available-rooms":{"get":{"tags":["search-controller"],"operationId":"searchAvailableRooms","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}}}}},"/api/v1/rooms/available":{"get":{"tags":["room-controller"],"operationId":"findAvailableRooms","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date-time"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Room"}}}}}}}},"/api/v1/reservations/customer/{customerId}":{"get":{"tags":["reservation-controller"],"operationId":"getReservationsByCustomerId","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/reservations/all":{"get":{"tags":["reservation-controller"],"operationId":"getAllReservations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/housekeeping-tasks/employee/{employeeId}":{"get":{"tags":["housekeeping-task-controller"],"operationId":"getTasksByEmployeeId","parameters":[{"name":"employeeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HousekeepingTaskResponse"}}}}}}}},"/api/v1/admin/{id}":{"get":{"tags":["admin-controller"],"operationId":"getAdmin","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}}}}}},"/api/v1/admin/employee/{username}":{"get":{"tags":["admin-controller"],"operationId":"getOneEmployee","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}}}}}},"/api/v1/admin/all":{"get":{"tags":["admin-controller"],"operationId":"getAllAdmins","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminDTO"}}}}}}}},"/api/v1/reservations/deleteRoomReservation/{reservationId}/{roomId}":{"delete":{"tags":["reservation-controller"],"operationId":"deleteReservationForOneRoom","parameters":[{"name":"reservationId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"roomId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/v1/reservations/delete/{id}":{"delete":{"tags":["reservation-controller"],"operationId":"deleteReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/delete/{id}":{"delete":{"tags":["admin-controller"],"operationId":"deleteEmployee","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminDTO"}}}}}}}},"components":{"schemas":{"UpdateRoomRequest":{"type":"object","properties":{"type":{"type":"string"},"status":{"type":"string"},"price":{"type":"number","format":"double"},"capacity":{"type":"integer","format":"int32"},"features":{"type":"string"}}},"RoomResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string"},"price":{"type":"number","format":"double"},"capacity":{"type":"integer","format":"int32"},"features":{"type":"string"}}},"Customer":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"enabled":{"type":"boolean"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"Invoice":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"number","format":"double"},"reservations":{"type":"array","items":{"$ref":"#/components/schemas/Reservation"}}}},"Reservation":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"status":{"type":"string"},"customer":{"$ref":"#/components/schemas/Customer"},"rooms":{"type":"array","items":{"$ref":"#/components/schemas/Room"}},"invoice":{"$ref":"#/components/schemas/Invoice"}}},"Room":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"status":{"type":"string"},"price":{"type":"number","format":"double"},"capacity":{"type":"integer","format":"int32"},"size":{"type":"string"},"facilities":{"type":"string"},"features":{"type":"string"}}},"Token":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"token":{"type":"string"},"tokenType":{"type":"string","enum":["BEARER"]},"revoked":{"type":"boolean"},"expired":{"type":"boolean"},"user":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","CUSTOMER"]},"tokens":{"type":"array","items":{"$ref":"#/components/schemas/Token"}},"enabled":{"type":"boolean"},"username":{"type":"string"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"InvoiceRequest":{"type":"object","properties":{"amount":{"type":"number","format":"double"},"reservationIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"InvoiceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"amount":{"type":"number","format":"double"},"reservationIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"HousekeepingTaskRequest":{"type":"object","properties":{"description":{"type":"string"},"status":{"type":"string"},"assignedEmployeeId":{"type":"integer","format":"int64"}}},"HousekeepingTaskResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"description":{"type":"string"},"status":{"type":"string"},"assignedEmployeeId":{"type":"integer","format":"int64"},"assignedEmployeeName":{"type":"string"}}},"Employee":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"salary":{"type":"number","format":"double"},"nationalNumber":{"type":"string"}}},"UpdateCustomerRequest":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"CustomerResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"}}},"AdminDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"salary":{"type":"number","format":"double"},"role":{"type":"string"},"age":{"type":"integer","format":"int32"}}},"CustomerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"ReservationRequest":{"type":"object","properties":{"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"status":{"type":"string"},"customer":{"$ref":"#/components/schemas/CustomerDTO"},"room":{"$ref":"#/components/schemas/RoomDTO"}}},"RoomDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"}}},"ChangePasswordRequest":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"},"confirmationPassword":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","CUSTOMER"]}}},"AuthenticationResponse":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}},"AuthenticationRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}